Fix version make target when not in git repo
When a git repo cannot be detected, the version target probably ran already
(e.g. when sources are distributed for packaging). In this case, the version target
should do nothing, exept to copy the version header file to the calling directory -
the copy allows invoking the Makefile from other directories.