Include setup.py for rtd versioning
Readthedocs does not allow the version fields to be filled when importing via the
github webhook - there is no make hook as in copr; it simply finds and copies the
relevant doc files. It does, however, allow the invocation of a setup.py script,
which we exploit here to populate the version fields. Note that this is somewhat
of a hack, and it should be revised if rtd decides to introduce relevant
infrastructure in the future.