Fix port hotplugging with DPDK 17.11
DPDK's --no-pci behaviour changed in 17.11. With the flag passed, the EAL never
populates its device lists which causes subsequent hotplugs to fail. When it is
not passed, EAL attaches to all uio devices, which we do not want that either.
The 'solution' is to omit the flag in order for DPDK to discover devices and to
also add a bogus whitelist address such that DPDK will only auto-attach to it.