Add disto to debian versioning string
PPA does not allow a shared package version between distros. The ubuntu version
is now added when preparing the debian changelog file during deb building. A
distro name <-> version number map is maintained in the changelog file to ease
automation and limit errors. To create a changelog for a custom distro, set
'UBUNTU_DISTRO=<distro>' when invoking the deb target and ensure that the
distro is in the provided map. The default distro is xenial.