Allocate mempools dynamically
Mempool freeing enables the dynamic allocation and freeing of mempools for relay
instances. With this feature, only mbuf pools which are in operation are
allocated at any give time. Dynamic allocation also aids in the creation of mbuf
pools of different sizes, which is needed to cater for bonded devices.