Robustify memory allocation upon VM connection
With dynamic memory allocation we allocate mempools for relay instances upon
VM/VF connection. For the VM case, we cannot assume the availability of memory
on the VM's socket, since memory allocation is based on the affinity of the
worker CPUs. To provide for such events, an attempt is made to allocate memory
on SOCKET_ID_ANY if the allocation on the VM's socket failed.