Set relay ID in the relays initialization function
The relay ID determines its mempool name. When the ID is only set upon VM
connection and the VF is added first, the mempool will be allocated with the
unset ID which will probably occupy the name for relay 0. This will cause
subsequent allocations for relay 0 to fail. The fix sets the ID in the relays
initialization function.