Rework vhostuser socket handling for dynamic sockets
Change the way in which vhostuser sockets are registered/deregistered in
preparation for dynamic sockets at arbitrary paths. An option is introduced to
switch between the old default, in which virtio-forwarder creates template
sockets at a predefined location, and the dynamic implementation which allows
for arbitarary sockets. Dynamic sockets facilitate Openstack plugging
integration.