Lock when freeing packets on VF/virtio remove
The VF and virtio remove functions are called with only the vio and dpdk locks
respectively. However, these functions may free packets used by the rx and tx
threads, and therefore we need to acquire the dpdk lock in worker_remove_vf and
the vio lock in worker_remove_virtio.