Mark tx_q_bitmap as volatile
A VM suspend/resume cycle may cause the VM to VF thread to hang in an infinite
loop due to a zeroed tx_q_bitmap. The vring state change notification which must
update the bitmap can arrive after the new device callback has started the VM to
VF packet processing thread. Optimizations allow this thread to use the zeroed
bitmap value, which is never updated because the variable is non-volatile.
Marking the bitmap as volatile enables the state change to propagate to the
worker thread.

NOTE: Old qemu versions (2.5) do not issue a vring state change following VM
resume, which leaves the tx_q_bitmap zeroed out following a suspend/resume
cycle and the underlying relay in a bad state.