Fix package release version to 1
The packages built from the make targets correspond to the first release at the
specific commit (this commit represents a unique version). The release fields in
the packaging subsystems should therefore be set as 1. The "BUILD" version
generated in the version make target now forms part of the version tag as it
should.