Fix live migration for DPDKs >= v17.11
Live migration with DPDK >= 17.11 triggers a segmentation fault when
rte_vhost_avail_entries is called, presumably at too late a stage of the
migration. The feature change callback is used to monitor VHOST_F_LOG_ALL, which
informs us that migration is being initiated. If so, we skip finding the
available entries and use the burst size instead.