Zero packet avail counters after dropping
The datapath drops TX and RX traffic when the VF and virtio ends are not
connected, respectively. When packets are dropped, the corresponding counter for
the number of available entriers is not zeroed accordingly. The mbuf pointers in
the internal buffers which pointed to these [now freed] packets are then never
substituted with valid content, unless perhaps worker_remove_(virtio|vf) resets
the counters. Live migration can trigger a scenario where the mempool for these
packets are freed while the invalid pointers remain. This has serious
consequences, such as low or no performance. This commit fixes the corruption by
appropriately zeroing the packet available counters after dropping.