Fix RX checksum offloading
Configuring PMD queues according to the advertised default rxconf does not
ensure that RX checksums are calculated in hardware. This commit changes queue
configuration to explicitely specify the desired RX checksums to offload in the
same way they are specified for the TX queue. In addition, port-level (in
contrast to queue-level) RX offloads are enabled through the hw_ip_checksum
field, since it was observed that queue-level offloads don't propagate for all
PMDs.