Create core scheduler socket by default
Creation of the core scheduler socket is dictated by the
VIO4WD_CORE_SCHED_ENABLE flag. Change the shipped config such that it is
enabled. The core scheduler service must still be started/enabled for it
to do its work. Debugging utilities like the load monitor and core
pinner will now work out of the box.