Fix invalid DPDK port message
Prior to this patch, when a device is removed from the relay the
dpdk_port field is set to -1. When a device is subsequently added to the
same relay the value of the dpdk_port field in the relay struct is still
set to -1. When rte_eth_dev_info_get() is called with
relay->dpdk.dpdk_port an "Invalid port_id" message is printed. This
patch passes port_id into rte_eth_dev_info_get. relay->dpdk.dpdk_port is
populated in configure_dev_with_virtio with port_id shortly after the
events of the rte_eth_dev_info_get() call that caused the error message.

Change-Id: I5e9aa2b286674a70f7a61e207a21e0f1a6852984
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>