Add vhost-user socket name and path to stats query
For debugging purposes it is useful to map a given relay to the
vhost-user socket it is currently servicing. This patch adds a
vhost-user socket field to the stats struct and the required zmq
infrastructure for this purpose. The virtioforwarder_stats.py script
will print this new field for relays that are currently servicing a
vhost-user socket. For relays that aren't associated with a vhost-user
socket yet the script will report "NONE"

Change-Id: I0a42303a291026479ccdd26c390ea676f2993e62
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>