Expand meson support to include packaging logic
The use of make is being deprecated in DPDK 20.11 and going
forward virtio-forwarder will also move towards using meson and ninja
to build the project. This commit adds the necessary meson
infrastructure to facilitate building rpm and debian packages. The
actual building logic is still hosted in the virtio-forwarder-dist repo.

The meson.build file will to a large extent mirror the current logic in
the Makefile

Change-Id: I4a18b8ad5f75bd5e4810a3ead88a192f83a8b9a9
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>