Only generate version file when needed
It is possible to build using meson outside of a git repository. For
example when building a virtio-forwarder rpm package using a previously
generated source rpm. If this is the case the build should not attempt
to generate the vrelay_version.h file as it will fail to grab the
correct version from the git metadata. When running meson from a source
rpm the vrelay_version.h file should have been copied into the srpm
along with the rest of the source code. If this is not the case meson
will error and exit accordingly

Fixes: 5ec5d88 ("Expand meson support to include packaging logic")
Change-Id: If04f7519aa127383ed98900784e00206ca5ba436
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>