Make vrelay_version.h logic more robust
During the compilation of virtio-forwarder a version.h file is created.
Depending on exactly how virtio-forwarder is built (locally vs Copr or
PPA) the version.h file will either be in the meson build root for local
builds or in the main root for source rpm/deb builds (Copr/PPA).

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>