CI: add statically linked package building
This commit adds a GitHub workflow to build statically linked packages
of virtio-forwarder. The virtio-forwarder binary contained in these
packages is statically linked to the DPDK libraries and the maximum
number of supported ethernet devices has been increased to 64.

Packages will always be built on a push event to the master branch
or any of the release branches but packages will by default only be
available as build artifacts. When a new virtio-forwarder release is
created this workflow will also additionally publish the build artifacts
as an official release. The release job is only triggered when the push
event also contains a git tag.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>