CI: set machine type to default when building DPDK
As part of building statically linked packages DPDK is also compiled
from source. Before this patch the default machine type of "native" was
used. This results in DPDK building with -march=native. This isn't very
portable and can cause issues on CPU's that don't have feature parity
with the build machine's CPU. This patch sets the machine type to
"default". On x86 this will conservatively set "-march=corei7" which is
a lot more portable than "native"

Older versions of DPDK did not export a -march flag in the pkgconfig
file. Virtio-forwarder will set the flag to core2. Change this to corei7
for consistency too.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>