Check for bond PMD availability before linking
Virtio-forwarder has to explicitly link to the bond PMD. This needs to
happen because the virtio-forwarder code contains calls from the bond
PMD but the PMD is normally linked in by means of the "plugin" method in
DPDK. The DPDK EAL init code will dlopen() all shared objects located in
RTE_EAL_PMD_PATH. The linker has no way of knowing that the bond object
will be linked in at run time with the plugin method hence the need to
explicitly link to it at compile time.

Not all versions of the dpdk-devel package ship with the bond PMD. This
commit adds a check before trying to blindly link against it.
Virtio-forwarder already has the needed #ifdef's to handle the case
where the bond PMD was not present at the time of building.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>