Relax socket pair removal requirements
Before this patch both the PCI address and the vhost-user socket path
had to be specified when removing a device from virtio-forwarder with
the virtioforwarder_port_control.py script. Since there is a one-to-one
mapping between relay, VF and vhost socket it is possible to identify
the PCI device to be removed given just the vhost socket. For example:

./virtioforwarder_port_control.py remove_sock \
				  --vhost-path=/tmp/virtiofwd1.sock

Removing the requirement to pass both the vhost socket and PCI address
makes it easier for certain plugging logic higher up in the stack to
remove devices from virtio-forwarder.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>