Add avx and sse4.2 to march flag when building statically
A previous commit changed architecture type form native to corei7. This
was done to address CPU instruction compatibility issues on older CPU's
however this change significantly impacted performance. This commit
adds AVX as an extension to corei7. AVX is added specifically to corei7
because some architectures like sandybridge and ivybridge are not known
by older GCC versions (GCC v4.8 shipped with CentOS 7 for example). SSE
is also added as it is a hard requirement for building DPDK.

Fixes: eeafcae1bcb2 ("CI: set machine type to default when building DPDK")

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>