Use L4 info when calculating hash with multi-queue ipv6
Prior to this change only the L3 source, destination and protocol type
was included in the ipv6 hash calculation used for distributing traffic
between the vhost queues. This commit adds the L4 source and destination
info to this hash calculation too. It is often the case that entropy on
incoming traffic is confined to L4 and this change allows traffic of
this nature to be spread across queues.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>