CI: Rework existing build test workflow
This commit adds a new github workflow called build-dynamically-linked.
It replaces the current workflow called "build" in functionality but
with a different design approach. The new workflow uses containers
"nativity" within the github workflow context to compile test
virtio-forwarder in various OS configurations instead of invoking docker
via a shell script and mounting the virtio-forwarder directory into it.

The other main difference between the two workflows is how DPDK is
consumed. The old workflow performed a source compilation and install
of DPDK to provide the libraries to link against. The new workflow will
just install the DPDK package available in upstream repositories and
compile against it. This approach saves some time and also mimics the
Copr and PPA workflows more closely.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>