Add ability to return data in a PortControlResponse message
This commit adds a string field to the PortControlResponse message that
can be used to return data from virtio-forwarder when requested by the
port control script. The string field is meant to be generic and future
"query" operations can make use of this field to return requested data
to the user.

Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Signed-off-by: Simon Horman <simon.horman@corigine.com>