Fix removing the port that leads to a crash
In the rte_eth_dev_close(), it will clear the port, so it can't get the
dev_info, dev_info maybe is NULL, so rte_dev_remove() maybe can leads the
crash.
Since DPDK v19.11, the return type of the rte_eth_dev_info_get() function
has changed from 'void' to 'int'. So we need to add an int variable to
receive the return value.

Signed-off-by: Peng Zhang <peng.zhang@corigine.com>
Signed-off-by: Zerun Fu <zerun.fu@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>